PN-15140 update fro RMA UOM changes.

+++ PAS_Database/dbo/Tables/Tables2/VendorCreditMemoDetail.sql
@@ -3,11 +3,11 @@
     [VendorCreditMemoId]       BIGINT          NOT NULL,
     [VendorRMADetailId]        BIGINT          NULL,
     [VendorRMAId]              BIGINT          NULL,
-    [Qty]                      BIGINT          NULL,
+    [Qty]                      DECIMAL (18, 6) NULL,
     [Notes]                    VARCHAR (MAX)   NULL,
-    [OriginalAmt]              DECIMAL (18, 2) NULL,
-    [ApplierdAmt]              DECIMAL (18, 2) NULL,
-    [RefundAmt]                DECIMAL (18, 2) NULL,
+    [OriginalAmt]              DECIMAL (18, 6) NULL,
+    [ApplierdAmt]              DECIMAL (18, 6) NULL,
+    [RefundAmt]                DECIMAL (18, 6) NULL,
     [RefundDate]               DATETIME2 (7)   NOT NULL,
     [MasterCompanyId]          INT             NOT NULL,
     [CreatedBy]                VARCHAR (256)   NOT NULL,
@@ -16,10 +16,8 @@
     [UpdatedDate]              DATETIME2 (7)   CONSTRAINT [DF_VendorCreditMemoDetail_UpdatedDate] DEFAULT (getutcdate()) NOT NULL,
     [IsActive]                 BIT             CONSTRAINT [DF_VendorCreditMemoDetail_IsActive] DEFAULT ((1)) NOT NULL,
     [IsDeleted]                BIT             CONSTRAINT [DF_VendorCreditMemoDetail_IsDeleted] DEFAULT ((0)) NOT NULL,
-    [UnitCost]                 DECIMAL (18, 2) NULL,
+    [UnitCost]                 DECIMAL (18, 6) NULL,
     [StockLineId]              BIGINT          NULL,
     CONSTRAINT [PK_VendorCreditMemoDetail] PRIMARY KEY CLUSTERED ([VendorCreditMemoDetailId] ASC)
 );
-
-
